refactor(ming-tts): streamline
