--- sglang_omni/models/ming_tts/audio_vae/__init__.py
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""Ming-Omni-TTS AudioVAE config package.
-
-The runtime AudioVAE implementation is shared with Ming-Omni talker.
-"""
-
-__all__: list[str] = []
