--- sglang_omni/models/ming_tts/fm/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""Ming-Omni-TTS flow-matching heads.
-
-The submodules are intentionally imported lazily by callers so package import
-does not require torch or x-transformers.
-"""
-
-__all__: list[str] = []
