+++ tests/unit_test/pipeline/test_scheduler.py
@@ -1075,36 +1075,3 @@ def test_omni_scheduler_result_adapter_failure_emits_error_without_raise() -> No
     assert scheduler._prefill_start_done == set()
     assert request_data.prefill_input_embeds is None
     assert request_data.decode_input_embeds is None
-
-
-def test_omni_scheduler_follower_stream_output_bypasses_result_adapter() -> None:
-    """TP followers clean local tensors without emitting external results."""
-    scheduler = object.__new__(OmniScheduler)
-    scheduler.outbox = Queue()
-    scheduler.is_entry_rank = False
-    scheduler._first_emit_done = {"req-follower"}
-    scheduler._prefill_start_done = {"req-follower"}
-
-    def fail_adapter(_data):
-        raise AssertionError("follower must not serialize final payload")
-
-    scheduler._result_adapter = fail_adapter
-    request_data = SimpleNamespace(
-        prefill_input_embeds=torch.ones(1),
-        decode_input_embeds=[torch.ones(1)],
-    )
-    req = SimpleNamespace(
-        rid="req-follower",
-        _omni_data=request_data,
-        output_ids=[1, 2],
-        finished=lambda: True,
-        finished_reason=None,
-    )
-
-    scheduler.stream_output([req])
-
-    assert scheduler.outbox.empty()
-    assert scheduler._first_emit_done == set()
-    assert scheduler._prefill_start_done == set()
-    assert request_data.prefill_input_embeds is None
-    assert request_data.decode_input_embeds is None
